Compiled with extra warnings for GCC \  -Wall -Wextra -Wpedantic -Werror

+++ Src/Scripts/Objects/video/RecordFrames.cpp
@@ -74,7 +74,7 @@ void RecordFrames::saveImageToFile(CRLCameraDataType type, const std::string &pa
             auto *d = (uint16_t *) ptr->data;
             std::vector<uint8_t> buf;
             buf.reserve(ptr->m_Width * ptr->m_Height);
-            for (int i = 0; i < ptr->m_Width * ptr->m_Height; ++i) {
+            for (size_t i = 0; i < ptr->m_Width * ptr->m_Height; ++i) {
                 d[i] /= 16;
                 uint8_t lsb = d[i] & 0x000000FF;
                 buf.emplace_back(lsb);
@@ -105,7 +105,7 @@ void RecordFrames::saveImageToFile(CRLCameraDataType type, const std::string &pa
 
             std::memcpy(src->data[0], ptr->data, ptr->m_Len);
             auto* d = (uint16_t*) ptr->data2;
-            for (int i = 0; i < ptr->m_Height / 2 * ptr->m_Width/2; ++i) {
+            for (size_t i = 0; i < ptr->m_Height / 2 * ptr->m_Width/2; ++i) {
                 src->data[1][i] = d[i] & 0xff;
                 src->data[2][i] = (d[i] >> (8)) & 0xff;
             }
